+++ tests/Integration/Servers/ServerTest.php
@@ -10,6 +10,7 @@ namespace Tests\Integration\Servers;
 
 use LKDev\HetznerCloud\Models\Images\Image;
 use LKDev\HetznerCloud\Models\ISOs\ISO;
+use LKDev\HetznerCloud\Models\Networks\Network;
 use LKDev\HetznerCloud\Models\Servers\Server;
 use LKDev\HetznerCloud\Models\Servers\Servers;
 use LKDev\HetznerCloud\Models\Servers\Types\ServerType;
@@ -201,5 +202,45 @@ class ServerTest extends TestCase
         $apiResponse = $this->server->metrics('cpu,disk,network', date("c"), date("c"), 60);
         $metrics = $apiResponse->getResponsePart('metrics');
         $this->assertEquals([["1435781470.622", "42"]], $metrics->time_series->name_of_timeseries->values ?? null);
+    }
+
+    public function testAttachToNetworkBasic(){
+        $apiResponse = $this->server->attachToNetwork(new Network(4711));
+        $this->assertEquals('attach_to_network', $apiResponse->getResponsePart('action')->command);
+        $this->assertEquals($this->server->id, $apiResponse->getResponsePart('action')->resources[0]->id);
+        $this->assertEquals('server', $apiResponse->getResponsePart('action')->resources[0]->type);
+
+        $this->assertEquals(4711, $apiResponse->getResponsePart('action')->resources[1]->id);
+        $this->assertEquals('network', $apiResponse->getResponsePart('action')->resources[1]->type);
+    }
+
+    public function testAttachToNetworkAdvanced(){
+        $apiResponse = $this->server->attachToNetwork(new Network(4711), "10.0.1.1", ["10.0.1.2"]);
+        $this->assertEquals('attach_to_network', $apiResponse->getResponsePart('action')->command);
+        $this->assertEquals($this->server->id, $apiResponse->getResponsePart('action')->resources[0]->id);
+        $this->assertEquals('server', $apiResponse->getResponsePart('action')->resources[0]->type);
+
+        $this->assertEquals(4711, $apiResponse->getResponsePart('action')->resources[1]->id);
+        $this->assertEquals('network', $apiResponse->getResponsePart('action')->resources[1]->type);
+    }
+
+    public function testDetachFromNetwork(){
+        $apiResponse = $this->server->detachFromNetwork(new Network(4711));
+        $this->assertEquals('detach_from_network', $apiResponse->getResponsePart('action')->command);
+        $this->assertEquals($this->server->id, $apiResponse->getResponsePart('action')->resources[0]->id);
+        $this->assertEquals('server', $apiResponse->getResponsePart('action')->resources[0]->type);
+
+        $this->assertEquals(4711, $apiResponse->getResponsePart('action')->resources[1]->id);
+        $this->assertEquals('network', $apiResponse->getResponsePart('action')->resources[1]->type);
+    }
+
+    public function testChangeAliasIPs(){
+        $apiResponse = $this->server->changeAliasIPs(new Network(4711),  ["10.0.1.2"]);
+        $this->assertEquals('change_alias_ips', $apiResponse->getResponsePart('action')->command);
+        $this->assertEquals($this->server->id, $apiResponse->getResponsePart('action')->resources[0]->id);
+        $this->assertEquals('server', $apiResponse->getResponsePart('action')->resources[0]->type);
+
+        $this->assertEquals(4711, $apiResponse->getResponsePart('action')->resources[1]->id);
+        $this->assertEquals('network', $apiResponse->getResponsePart('action')->resources[1]->type);
     }
 }