+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 1.7.0
+ + Add `Networks` support ( `LKDev\HetznerCloud\Models\Networks\Networks` & `LKDev\HetznerCloud\Models\Networks\Network`)
+ + Add `networkZone` property to `LKDev\HetznerCloud\Models\Locations\Location`
+
 ## 1.6.2 (27.05.2019) 
  + Add `volumes` and `automount` parameters to `LKDev\HetznerCloud\Models\Servers\Servers` - `createInDatacenter` and `createInLocation`
  + Add `created` property to `LKDev\HetznerCloud\Models\Servers\Server`